+++ app/(resume)/cv/page.tsx
@@ -5,7 +5,7 @@ export const metadata = {
 
 import Hero from '@/components/hero'
 import Features from '@/components/features'
-import Newsletter from '@/components/newsletter'
+import Contact from '@/components/contact'
 import Zigzag from '@/components/zigzag'
 import Testimonials from '@/components/testimonials'
 
@@ -14,7 +14,7 @@ export default function Resume() {
     <>
       <Zigzag />
       <Testimonials />
-      <Newsletter />
+      <Contact />
     </>
   )
 }
